requirements/requirements.txt
appnope==0.1.3
asttokens==2.2.1
Babel==2.12.1
backcall==0.2.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
ipykernel==6.23.0
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.0
plotly==5.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
tenacity==8.2.2
tkcalendar==1.6.1
tornado==6.3.1
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6